fix(docker): Close #7
The docker client API with json stream doesn't handle error message,
which seems perfectly appropriate.

The decoder is now synchronous, which remove a lot of complexity with the
previous pipeline mechanism.

Also, if the decoder parse an error, it will be returned and handled correctly
by both the docker client api and this application.